feat: extract shared Storefront component + public /store/[slug] page
Extracts the global storefront's body into components/store/Storefront.tsx
so it can be reused as the base template for per-store storefronts. The
global /store page is a pure refactor with no behavior change. A new
/store/[slug] page renders the same component scoped to a store's curated
catalog via the existing /api/v1/stores/[slug]/{featured,search} routes,
and 404s cleanly on an unknown slug. StoreTopNav/StoreLayout now accept an
optional store context to scope branding, nav links, and search.

Closes #140.
